增加自定义类型转换器

+++ src/main/java/com/jonas/dao/domain/User.java
@@ -2,13 +2,15 @@ package com.jonas.dao.domain;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class User {
     private Integer id;
     private String name;
     private Integer age;
     private Integer status;
-    private Integer score;
+    private BigInteger score;
     private Long ctime;
     private Long utime;
 }
